refactor(kernel): extract the axis scope policy into its own module (BI-AA7D80FE)
The Module Size Guard was right: wiki-taxonomy.ts hit 819 LOC against the
800 ceiling. The split it forced is the better structure anyway —
wiki-taxonomy.ts owns the axis REGISTRY ("what axes exist"), and the new
dimension-scope.ts owns the axis SCOPE POLICY ("which are shared doctrine and
which belong to a profession").

No behaviour change: the classification, projections, integrity assertion and
tests move verbatim. Adds the `@dpf/db/dimension-scope` export so apps/web
imports the policy directly rather than through the registry module.
