+++ packages/db/package.json
@@ -42,6 +42,7 @@
     "./seed-deliberation": "./src/seed-deliberation.ts",
     "./edge-node-types": "./src/edge-node-types.ts",
     "./wiki-taxonomy": "./src/wiki-taxonomy.ts",
+    "./dimension-scope": "./src/dimension-scope.ts",
     "./wiki-store": "./src/wiki-store.ts",
     "./wiki-frontmatter": "./src/wiki-frontmatter.ts",
     "./capability-maturity": "./src/capability-maturity.ts",
